Created characters folder \w test imgs and classes file with player
the player is still to be refined since will handle the most of the logic as collision and so on actually is just displaying.
The 2 images are there just for testing. I commented this not much well so if you need some help understanding code feel free to ask is midnight here and i got to go to bed sry